Completes the E2E test

+++ endpoints/delete_endpoint.py
@@ -1,4 +1,6 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
+from fastapi.responses import RedirectResponse
+from functions.fetch_url_details import fetch_url_details
 from functions.get_db import get_db
 
 router = APIRouter()
@@ -6,8 +8,11 @@ router = APIRouter()
 @router.delete("/delete")
 def delete_url(alias_id: str):
     # TODO: Validation + auth
+    if fetch_url_details(alias_id) is None:
+        raise HTTPException(status_code=404, detail="Alias not found")
+
     (con,cur) = get_db()
     cur.execute(f"DELETE FROM urls WHERE alias = '{alias_id}'")
     con.commit()
-    return {"alias": alias_id}
+    return {"detail": "Alias deleted"}
